Refactor: Separate types from assertions
